refactor: 独立出HResultException类以及一些通用函数
