+++ TrafficMonitor/D2D1Support.cpp
@@ -1,89 +1,13 @@
 ﻿#include "stdafx.h"
 #include "D2D1Support.h"
-#include "Common.h"
-#include "TrafficMonitor.h"
 
 #pragma comment(lib, "D2d1.lib")
 #pragma comment(lib, "Dwrite.lib")
-
-namespace FunctionChecker
-{
-    namespace Details
-    {
-        template <class Strategy>
-        bool StrategicCheck(LPCTSTR p_library_name, Strategy strategy)
-        {
-            bool result = false;
-            auto hmodule = ::LoadLibrary(p_library_name);
-            if (hmodule)
-            {
-                result = strategy(hmodule);
-                ::FreeLibrary(hmodule);
-            }
-            return result;
-        }
-    }
-} // namespace FunctionChecker
-
-bool FunctionChecker::CheckLibraryExist(LPCTSTR p_library_name) noexcept
-{
-    return Details::StrategicCheck(p_library_name,
-                                   [](HMODULE)
-                                   { return true; });
-}
-
-bool FunctionChecker::CheckFunctionExist(LPCTSTR p_library_name, LPCSTR p_function_name) noexcept
-{
-    return Details::StrategicCheck(p_library_name,
-                                   [p_function_name](HMODULE h_library) -> bool
-                                   {
-                                       auto* p_test_function = ::GetProcAddress(h_library, p_function_name);
-                                       return p_test_function != NULL;
-                                   });
-}
-
-void ThrowIfFailed(HRESULT hr, const char* p_message)
-{
-    ThrowIfFailed<HResultException>(std::move(hr), std::move(p_message));
-}
-
-HResultException::HResultException(HRESULT hr, const char* p_message)
-    : std::runtime_error{p_message}
-{
-    m_hr = ::GetErrorInfo(0, &m_p_error);
-}
-
-auto HResultException::GetError()
-    -> Microsoft::WRL::ComPtr<IErrorInfo>
-{
-    return m_p_error;
-}
-
-bool HResultException::HasError() const noexcept
-{
-    return SUCCEEDED(m_hr);
-}
 
 bool D2D1Support::CheckSupport()
 {
     static bool result = FunctionChecker::CheckFunctionExist(_T("D2d1.dll"), "D2D1CreateFactory");
     return result;
-}
-
-void LogHResultException(HResultException& ex)
-{
-    auto* log = ex.what();
-    //写入日志
-    CCommon::WriteLog(log, theApp.m_log_path.c_str());
-    auto* p_error = ex.GetError().Get();
-    if (p_error == NULL)
-    {
-        return;
-    }
-    BSTR p_description{NULL};
-    ThrowIfFailed(p_error->GetDescription(&p_description));
-    CCommon::WriteLog(p_description, theApp.m_log_path.c_str());
-    ::SysFreeString(p_description);
 }
 
 ID2D1Factory* D2D1Support::GetFactory()
