+++ TrafficMonitor/D2D1Support.h
@@ -1,38 +1,8 @@
 ﻿#pragma once
 #include <stdexcept>
-#include <wrl/client.h> //ComPtr
 #include <d2d1.h>
 #include <dwrite.h>
-
-#define RELEASE_COM(p)     \
-    {                       \
-        if (p)              \
-        {                   \
-            (p)->Release(); \
-            (p) = (NULL);   \
-        }                   \
-    }
-
-namespace FunctionChecker
-{
-    bool CheckLibraryExist(LPCTSTR p_library_name) noexcept;
-    bool CheckFunctionExist(LPCTSTR p_library_name, LPCSTR p_function_name) noexcept;
-}
-
-class HResultException : public std::runtime_error
-{
-public:
-    HResultException(HRESULT hr, const char* p_message);
-    ~HResultException() override = default;
-
-    auto GetError()
-        -> Microsoft::WRL::ComPtr<IErrorInfo>;
-    bool HasError() const noexcept;
-
-private:
-    Microsoft::WRL::ComPtr<IErrorInfo> m_p_error;
-    HRESULT m_hr;
-};
+#include "HResultException.h"
 
 class D2D1Exception final : public HResultException
 {
@@ -43,20 +13,6 @@ class DWriteException final : public HResultException
 {
     using HResultException::HResultException;
 };
-
-#define D2D1_SUPPORT_DEFAULT_EXPECTION_TIP "Error occurred when call COM function."
-void ThrowIfFailed(HRESULT hr, const char* p_message = D2D1_SUPPORT_DEFAULT_EXPECTION_TIP);
-template <class Exception>
-void ThrowIfFailed(HRESULT hr, const char* p_message = D2D1_SUPPORT_DEFAULT_EXPECTION_TIP)
-{
-    if (FAILED(hr))
-    {
-        throw Exception{hr, p_message};
-    }
-}
-#undef D2D1_SUPPORT_DEFAULT_EXPECTION_TIP
-
-void LogHResultException(HResultException& ex);
 
 class D2D1Support
 {
