+++ TrafficMonitor/D3D10Support1.cpp
@@ -1,0 +1,2 @@
+#include "stdafx.h"
+#include "D3D10Support1.h"
