+++ TrafficMonitor/D3D10Support1.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <list>
+#include <d3d10_1.h>
+#include "HResultException.h"
+
