+++ TrafficMonitor/TrafficMonitor.vcxproj
@@ -514,13 +514,16 @@
     <ClInclude Include="CTabCtrlEx.h" />
     <ClInclude Include="CVariant.h" />
     <ClInclude Include="D2D1Support.h" />
+    <ClInclude Include="D3D10Support1.h" />
     <ClInclude Include="DisplayTextSettingDlg.h" />
     <ClInclude Include="DrawCommonEx.h" />
+    <ClInclude Include="DrawTextManager.h" />
     <ClInclude Include="FilePathHelper.h" />
     <ClInclude Include="HighResolutionTimer.h" />
     <ClInclude Include="HistoryTrafficCalendarDlg.h" />
     <ClInclude Include="HistoryTrafficFile.h" />
     <ClInclude Include="HistoryTrafficListDlg.h" />
+    <ClInclude Include="HResultException.h" />
     <ClInclude Include="LinkStatic.h" />
     <ClInclude Include="ColorStatic.h" />
     <ClInclude Include="Common.h" />
@@ -587,12 +590,15 @@
     <ClCompile Include="CTabCtrlEx.cpp" />
     <ClCompile Include="CVariant.cpp" />
     <ClCompile Include="D2D1Support.cpp" />
+    <ClCompile Include="D3D10Support1.cpp" />
     <ClCompile Include="DisplayTextSettingDlg.cpp" />
     <ClCompile Include="DrawCommonEx.cpp" />
+    <ClCompile Include="DrawTextManager.cpp" />
     <ClCompile Include="FilePathHelper.cpp" />
     <ClCompile Include="HistoryTrafficCalendarDlg.cpp" />
     <ClCompile Include="HistoryTrafficFile.cpp" />
     <ClCompile Include="HistoryTrafficListDlg.cpp" />
+    <ClCompile Include="HResultException.cpp" />
     <ClCompile Include="LinkStatic.cpp" />
     <ClCompile Include="ColorStatic.cpp" />
     <ClCompile Include="Common.cpp" />
