+++ TrafficMonitor/TrafficMonitor.vcxproj.filters
@@ -199,6 +199,15 @@
     <Filter Include="源文件和头文件\公共的类\Nullable">
       <UniqueIdentifier>{f51d70bc-ec1d-45e9-9c57-777da2b6e51d}</UniqueIdentifier>
     </Filter>
+    <Filter Include="源文件和头文件\公共的类\DrawTextManager">
+      <UniqueIdentifier>{af0ec00c-027f-4003-890e-7122d0b8df0c}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="源文件和头文件\公共的类\HResultException">
+      <UniqueIdentifier>{48a72bdd-aca1-4367-9907-4bbb92dea212}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="源文件和头文件\公共的类\D3D10Support1">
+      <UniqueIdentifier>{f6a75faf-64b5-4f3b-a382-7dfd4a17d61c}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <Text Include="ReadMe.txt" />
@@ -423,7 +432,13 @@
       <Filter>源文件和头文件\公共的类\Nullable</Filter>
     </ClInclude>
     <ClInclude Include="DrawTextManager.h">
-      <Filter>源文件和头文件\公共的类\DrawCommon</Filter>
+      <Filter>源文件和头文件\公共的类\DrawTextManager</Filter>
+    </ClInclude>
+    <ClInclude Include="HResultException.h">
+      <Filter>源文件和头文件\公共的类\HResultException</Filter>
+    </ClInclude>
+    <ClInclude Include="D3D10Support1.h">
+      <Filter>源文件和头文件\公共的类\D3D10Support1</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -626,7 +641,13 @@
       <Filter>源文件和头文件\公共的类\D2D1Support</Filter>
     </ClCompile>
     <ClCompile Include="DrawTextManager.cpp">
-      <Filter>源文件和头文件\公共的类\DrawCommon</Filter>
+      <Filter>源文件和头文件\公共的类\DrawTextManager</Filter>
+    </ClCompile>
+    <ClCompile Include="HResultException.cpp">
+      <Filter>源文件和头文件\公共的类\HResultException</Filter>
+    </ClCompile>
+    <ClCompile Include="D3D10Support1.cpp">
+      <Filter>源文件和头文件\公共的类\D3D10Support1</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
